LayoutTests/platform/ios-simulator/http/tests/media/hls/video-controls-live-stream-expected.txt


EVENT(play)
EXPECTED (video.duration == 'Infinity') OK
-webkit-media-text-track-container: classes: [hidden]
-webkit-media-show-controls: classes: []
-webkit-media-controls-start-playback-button: classes: []
-webkit-media-controls-start-playback-background: classes: [webkit-media-controls-start-playback-background]
-webkit-media-controls-start-playback-tint: classes: []
-webkit-media-controls-start-playback-glyph: classes: [webkit-media-controls-start-playback-glyph]
TEST(timeline == null || timeline.disabled) OK
END OF TEST
